queue/requirements.txt
autopep8==2.0.4
certifi==2023.7.22
charset-normalizer==3.3.1
confluent-kafka==2.2.0
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
dependency-injector==4.41.0
fonttools==4.43.1
greenlet==3.0.0
idna==3.4
kiwisolver==1.4.5
lxml==4.9.3
matplotlib==3.8.0
numpy==1.26.1
obspy==1.4.0
packaging==23.2
Pillow==10.1.0
pycodestyle==2.11.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
scipy==1.11.3
six==1.16.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
urllib3==2.0.7
prometheus_client
psutil
